requirements.txt
django>=3.0.7,<3.1.0
# Python==3.7.0
# Optional packages which may be used with REST framework.
psycopg2-binary>=2.8.5, <2.9
markdown==3.1.1
pygments==2.4.2
django-guardian==2.2.0
django-filter>=2.2.0, <2.3
coreapi==2.3.1
coreschema==0.0.4
pyyaml>=5.1
djangorestframework==3.11.0
# Wheel for PyPI installs.
wheel==0.34.2

# channels
dj-rest-auth==1.1.2
djangorestframework-simplejwt==4.4.0
pyjwt==1.7.1
django-elasticsearch-dsl==7.1.4
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
python-dateutil==2.8.1
Automat==20.2.0
PyHamcrest==2.0.2
autobahn==20.7.1
cffi==1.14.3
channels==2.4.0
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
hyperlink==20.0.1
incremental==17.5.0
pyasn1-modules==0.2.8
pycparser==2.20
pyopenssl==19.1.0
service-identity==18.1.0
twisted==20.3.0
txaio==20.4.1
zope.interface==5.1.2
ted==20.3.0
txaio==20.4.1
zope.interface==5.1.2

# Twine for secured PyPI uploads.
twine==3.1.1

# Transifex client for managing translation resources.
transifex-client==0.13.9




# PEP8 code linting, which we run on all commits.
flake8==3.7.9
flake8-tidy-imports==4.1.0
pycodestyle==2.5.0

# Sort and lint imports
isort==4.3.21




# MkDocs to build our documentation.
mkdocs==1.1


# Pytest for running the tests.
pytest>=5.4.1,<5.5
pytest-django>=3.9.0,<3.10
pytest-cov>=2.7.1
